requirements.txt
langchain==0.0.203
streamlit==1.23.1
streamlit-chat==0.0.2.2
python-dotenv==1.0.0
openai==0.27.8
pinecone-client==2.2.1
pypdf==3.9.1
google-search-results==2.4.2